增加工时管理

+++ pmlite/apis/task.py
@@ -3,7 +3,7 @@ from flask import Blueprint, request
 from flask_sqlalchemy.pagination import Pagination
 from flask_jwt_extended import current_user, jwt_required
 
-from ..models import TaskModel, TaskTypeModel
+from ..models import TaskModel, TaskTypeModel, ManHourModel
 from ..extensions import db
 
 task_api = Blueprint("task", __name__, url_prefix="/task")
@@ -79,7 +79,6 @@ def task_add():
     try:
         task.save()
     except Exception as e:
-        print(e)
         return {
             'code': -1,
             'msg': '新增数据失败'
@@ -112,10 +111,93 @@ def task_edit(tid):
 
 # 删除
 @task_api.delete('/<int:tid>')
+@jwt_required()
 def task_delete(tid):
     task: TaskModel = db.get_or_404(TaskModel, tid)
+    if task.creator.name == current_user.name:
+        try:
+            pass
+            db.session.delete(task)
+            # user.is_del = True
+            db.session.commit()
+        except Exception as e:
+            return {
+                'code': -1,
+                'msg': '删除数据失败'
+            }
+        return {
+            'code': 0,
+            'msg': '删除数据成功'
+        }
+    else:
+        return{
+            'code': -1,
+            'msg': "只能删除自己创建的任务。"
+        }
+
+
+# 获取某任务工时列表
+@task_api.get('/<int:tid>/man-hour')
+def man_hour_list(tid):
+    q = db.select(ManHourModel).where(ManHourModel.task_id == tid)
+    MHlist = db.session.execute(q).scalars().all()
+    return {
+        'code': 0,
+        'msg': '信息查询成功',
+        'count': "",
+        'data': [item.json() for item in MHlist]
+    }
+
+
+# 添加工时
+@task_api.post('/<int:tid>/man-hour')
+@jwt_required()
+def man_hour_add(tid):
+    data = request.get_json()
+    man_hour = ManHourModel()
+    man_hour.update(data)
+    man_hour.user_id = current_user.id
+    print(man_hour)
     try:
-        db.session.delete(task)
+        man_hour.save()
+    except Exception as e:
+        return {
+            'code': -1,
+            'msg': '新增数据失败'
+        }
+    return {
+        'code': 0,
+        'msg': '新增数据成功'
+    }
+
+
+# 修改工时
+@task_api.put('/<int:tid>/man-hour/<int:mid>')
+def man_hour_edit(tid, mid):
+    data = request.get_json()
+    man_hour = db.get_or_404(ManHourModel, mid)
+    man_hour.update(data)
+    try:
+        man_hour.save()
+    except Exception as e:
+        return {
+            'code': -1,
+            'msg': '修改数据失败'
+        }
+    return {
+        'code': 0,
+        'msg': '修改数据成功'
+    }
+
+
+# 删除
+@task_api.delete('/<int:tid>/man-hour/<int:mid>')
+@jwt_required()
+def man_hour_delete(tid, mid):
+    man_hour: ManHourModel = db.get_or_404(ManHourModel, mid)
+    try:
+        pass
+        db.session.delete(man_hour)
         # user.is_del = True
         db.session.commit()
     except Exception as e:
